scripts/requirements.txt
connect-cli
homebrew-pypi-poet
